Bisect on examples and reconstruct order tree
By implementing Node#filter, we can pass an array of leaf nodes to #filter and
receive a tree that reaches them all, which we can then use as the basis for
ordering the spec run.

This means we can bisect only on the leaf nodes, which in turn means we can
reach a solution faster (as the group nodes are ignored).